DEF-193: abstract observer to hook
